requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
sqlalchemy==2.0.36
psycopg2-binary==2.9.10
pydantic==2.10.3
xgboost==2.1.3
scikit-learn==1.6.0
pandas==2.2.3
numpy==2.2.0
python-dotenv==1.0.1
